+++ backend/src/scripts/run_dev_server.py
@@ -1,10 +1,13 @@
+import os
+
 from backend.main import create_app
 
 
 def main() -> None:
     """Run Flask development server."""
     app = create_app()
-    app.run(debug=True, port=5555)
+    testing = os.environ.get("FLASK_ENV") == "TESTING"
+    app.run(debug=True, port=5555, use_reloader=not testing)
 
 
 if __name__ == "__main__":